+++ ivy/functional/backends/numpy/activations.py
@@ -19,14 +19,6 @@ def relu(x: np.ndarray, /, *, out: Optional[np.ndarray] = None) -> np.ndarray:
 relu.support_native_out = True
 
 
-@_scalar_output_to_0d_array
-def relu6(x: np.ndarray, /, *, out: Optional[np.ndarray] = None) -> np.ndarray:
-    return np.minimum(np.maximum(x, 0, dtype=x.dtype), 6, out=out, dtype=x.dtype)
-
-
-relu6.support_native_out = True
-
-
 def leaky_relu(
     x: np.ndarray, /, *, alpha: float = 0.2, out: Optional[np.ndarray] = None
 ) -> np.ndarray:
