+++ ivy/functional/backends/numpy/experimental/activations.py
@@ -31,3 +31,11 @@ def thresholded_relu(
 
 
 thresholded_relu.support_native_out = True
+
+
+@_scalar_output_to_0d_array
+def relu6(x: np.ndarray, /, *, out: Optional[np.ndarray] = None) -> np.ndarray:
+    return np.minimum(np.maximum(x, 0, dtype=x.dtype), 6, out=out, dtype=x.dtype)
+
+
+relu6.support_native_out = True
