+++ ivy/functional/backends/tensorflow/activations.py
@@ -33,11 +33,6 @@ def relu(x: Tensor, /, *, out: Optional[Tensor] = None) -> Tensor:
     return tf.nn.relu(x)
 
 
-@with_unsupported_dtypes({"2.9.1 and below": ("complex",)}, backend_version)
-def relu6(x: Tensor, /, *, out: Optional[Tensor] = None) -> Tensor:
-    return tf.nn.relu6(x)
-
-
 def sigmoid(x: Tensor, /, *, out: Optional[Tensor] = None) -> Tensor:
     if not ivy.is_array(x):
         x = float(x)
