+++ ivy/functional/backends/torch/activations.py
@@ -20,11 +20,6 @@ def relu(x: torch.Tensor, /, *, out: Optional[torch.Tensor] = None) -> torch.Ten
 
 
 @with_unsupported_dtypes({"1.11.0 and below": ("complex", "float16")}, backend_version)
-def relu6(x: torch.Tensor, /, *, out: Optional[torch.Tensor] = None) -> torch.Tensor:
-    return torch.nn.functional.relu6(x)
-
-
-@with_unsupported_dtypes({"1.11.0 and below": ("complex", "float16")}, backend_version)
 def leaky_relu(
     x: torch.Tensor,
     /,
